requirements.txt
alembic==1.13.1
aniso8601==9.0.1
astroid==3.2.2
blinker==1.8.2
cachetools==5.3.3
catboost==1.2.5
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
dill==0.3.8
docker==7.1.0
entrypoints==0.4
flake8==7.1.0
Flask==3.0.3
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
graphviz==0.20.3
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
loguru==0.7.2
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
mlflow==2.14.1
numpy==1.26.4
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
protobuf==4.25.3
psycopg2==2.9.9
pyarrow==15.0.2
pycodestyle==2.12.0
pyflakes==3.2.0
pylint==3.2.3
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.1.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.4.1
threadpoolctl==3.5.0
tomlkit==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
